@@ -1,6 +1,6 @@
 {
   "name": "softwarepunt/instarecord",
-  "description": "ActiveRecord Database ORM layer for MySQL",
+  "description": "ActiveRecord Database ORM layer for MySQL and PostgreSQL",
   "type": "library",
   "license": "MIT",
   "authors": [
@@ -11,9 +11,13 @@
   ],
   "minimum-stability": "stable",
   "require": {
-    "php": ">=8.1",
+    "php": ">=8.2",
     "ext-pdo": "*",
     "ext-json": "*"
+  },
+  "suggest": {
+    "ext-pdo_mysql": "MySQL support",
+    "ext-pdo_pgsql": "PostgreSQL support"
   },
   "require-dev": {
     "phpunit/phpunit": "^9"
